% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMoE_cv.R
\name{calcCriterion}
\alias{calcCriterion}
\title{Evaluate fitted NEMoE obejct}
\usage{
calcCriterion(NEMoE_obj, crit_list = c("AIC", "BIC", "ICL1"), ...)
}
\arguments{
\item{crit_list}{a list of evaluation metric.}

\item{...}{other parameters can be put into calcCriterion.}

\item{NEMoE}{a NEMoE of object with fitted parameters.}
}
\value{
A vector of evaluation result.
}
\description{
This function calculate evaluation of NEMoE object.
}
\examples{
data(NEMoE_example)
calcCriterion(NEMoE_example)
}
\seealso{
\code{\link{createCVList}}
}
