% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcdf}
\alias{calcdf}
\title{Calculate degree of freedom}
\usage{
calcdf(NEMoE_obj, output = "all")
}
\arguments{
\item{NEMoE_obj}{A NEMoE object with fitted result.}

\item{output}{whether degree of freedom with combined
components or compute within each component.}
}
\value{
a list of degree of freedom in gating network and experts network.
}
\description{
This function calculate degree of freedom of NEMoE object.
}
\examples{
data(NEMoE_example)
calcdf(NEMoE_example)
}
