% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcLL}
\alias{calcLL}
\title{Cpp function: calculate the four type of log likelihood function
(observed likelihood, penalized likelihood,
complete likelihood, penalized complete likelihood)
in all levels.}
\usage{
calcLL(X, Z, y, seg, W, V, lambda1, lambda2, alpha1, alpha2, multiply = FALSE)
}
\arguments{
\item{X}{an aggregated data matrix (n*P) of input in experts network
(rbind of all levels input).}

\item{Z}{a data matrix (n*q) of input in gating network.}

\item{y}{a vector of response (n).}

\item{seg}{an integer vector of length in each level (L).}

\item{W}{aggregated parameters in experts network (cbind of all levels)
((P+L)*K).}

\item{V}{parameters in gating network((q+1)*K).}

\item{lambda1}{aggregated penalty lambda parameters in experts network(P*K).}

\item{lambda2}{penalty parameters lambda in gating network(q*1).}

\item{alpha1}{penalty parameters alpha in experts network(L*K).}

\item{alpha2}{penalty parameter alpha in gating network(1).}

\item{multiply}{a bool variable indicate whether lambda parameters
times sample size.}
}
\value{
A matrix of four types of likelihood function.
}
\description{
Cpp function: calculate the four type of log likelihood function
(observed likelihood, penalized likelihood,
complete likelihood, penalized complete likelihood)
in all levels.
}
