% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcProb}
\alias{calcProb}
\title{Cpp function: Calculate the estimated probability using softmax function
(multinomial regression).}
\usage{
calcProb(X, B)
}
\arguments{
\item{X}{a data matrix of input in multinomial regression.}

\item{B}{coefficients in multinomial regression.}
}
\value{
A matrix of multinomial probability with
p_i=exp(X \%\emph{\% B_i)/sum(X \%}\% B)
}
\description{
Cpp function: Calculate the estimated probability using softmax function
(multinomial regression).
}
