% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compLikeli}
\alias{compLikeli}
\title{Cpp function: Complete likelihood function of mixture distribution}
\usage{
compLikeli(X, Z, y, W, V)
}
\arguments{
\item{X}{a data matrix of input in experts network.}

\item{Z}{a data matrix of input in gating network.}

\item{y}{a vector of response.}

\item{W}{parameters in experts network.}

\item{V}{parameters in gating network.}
}
\value{
A matrix of complete likelihood by
compLikeli = pi * Likelihood = pi*(1/1+exp(-X*W))^y
pi = exp(Z * V_i)/sum(Z * V)
}
\description{
Cpp function: Complete likelihood function of mixture distribution
}
