% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compTransform}
\alias{compTransform}
\title{Transformation of compositional data}
\usage{
compTransform(X, eps = 1e-04, method = "comp", scale = F)
}
\arguments{
\item{X}{a matrix to transform.}

\item{eps}{when elements of X smaller than eps, will be eps to avoid invalid value in log transformation.}

\item{method}{The method of transformation.
If method = "asin" will perform arcsine transformation.
If method = "clr" will perform central log transformation.
If method = "comp" will use composition data.}

\item{scale}{a logical variable to determine whether scale the transformed data.}
}
\value{
A matrix of transformed data.
}
\description{
Transformation of compositional data
}
