% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{createParameterList}
\alias{createParameterList}
\title{Create list of parameters for fitting NEMoE}
\usage{
createParameterList(
  lambda1 = 0.02,
  lambda2 = 0.015,
  alpha1 = 0.5,
  alpha2 = 0.5,
  beta_max = 10,
  EM_alg = "EM",
  init = "kmeans",
  itmax = 100,
  itmin = 1,
  adapt = TRUE,
  btr = TRUE,
  stop_all = TRUE,
  verbose = TRUE,
  early_stop = FALSE
)
}
\arguments{
\item{lambda1}{Penalty regularizer for the experts.}

\item{lambda2}{Penalty regularizer for the gating network.}

\item{alpha1}{Elastic net penalty value for experts.}

\item{alpha2}{Elastic net penalty value for gating network.}

\item{beta_max}{Maximal of coefficient. By default is 10.}

\item{EM_alg}{Method for Expecation maximization update.
Can be chosen from "EM", "CEM", "GEM", "SEM", "SAEM", "GEM".
By default is "EM".}

\item{init}{Method for initialization.
Can be chosen from "rand", "kmeans" and "glmnet".
If init="rand" will use a dirichlet distribution initialing the latent class.
If init="kmeans" the latent class will initialized using kmeans
clustering of input for gating network.
If init="glmnet" will use a equal probability
with lasso as its corresponding coefficients in experts network.}

\item{itmax}{Maximium number of iteration in fitting NEMoE. By default is 100.}

\item{itmin}{Minimium number of iteration in fitting NEMoE. By default is 3.}

\item{adapt}{whether use adaptive mode in optimization. By default is TRUE.}

\item{btr}{Whether use backtracking during the iteration. By default is TRUE.}

\item{stop_all}{Method of stop criterion. If stop_all = TRUE means that either
coefficient or loss function converge. If stop_all = FALSE means that both
coefficient and loss function converge.}

\item{verbose}{A logical input indicating whether the intermediate
steps will be printed.}

\item{early_stop}{A logical input indicate whether early stop when one of
the fitted latent class have zero variables selected (to save time).
By default is TRUE.}
}
\value{
A list contain parameters in fitting NEMoE.
}
\description{
This function create parameters that put into fitNEMoE function.
}
\examples{
params = createParameterList(lambda1 = c(0.005, 0.01, 0.02, 0.025))
}
\seealso{
\code{\link{NEMoE_buildFromList}},
\code{\link{NEMoE_buildFromPhyloseq}}, \code{\link{fitNEMoE}}
}
