% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMoE_cv.R
\name{cvNEMoE}
\alias{cvNEMoE}
\title{Parameters tunning in NEMoE}
\usage{
cvNEMoE(NEMoE, verbose = T, ...)
}
\arguments{
\item{NEMoE}{a NEMoE of object without cross validated parameters.}

\item{verbose}{A logical input indicating whether the intermediate
steps will be printed.}

\item{...}{Other parameters that can be passed to cvNEMoE.}
}
\value{
A NEMoE object with different lambdas.
}
\description{
This function calculate evaluation of NEMoE object.
}
\examples{
data(NEMoE_example)
NEMoE_example = cvNEMoE(NEMoE_example, verbose = F)
}
\seealso{
\code{\link{createCVList}}
}
