% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cvtLabel}
\alias{cvtLabel}
\title{Convert label to factor of matrix}
\usage{
cvtLabel(prob, idx = FALSE)
}
\arguments{
\item{prob}{a matrix to convert.}

\item{idx}{the type of output. If idx= TRUE will
return the index of largest probability of each row. If idx= FALSE will return a one-hot matrix.}
}
\value{
a index of largest probability or one hot matrix.
}
\description{
This function convert a soft probability matrix to factor or one hot matrix.
}
