% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterComp}
\alias{filterComp}
\title{Filter the composition matrix}
\usage{
filterComp(X, thresh_func = var, thresh = 1e-04, id_out = FALSE)
}
\arguments{
\item{X}{the matrix input to filter}

\item{thresh_func}{a function to calculate statistics for filtering.
By default is variance.}

\item{thresh}{If the statistics of variable smaller than threshold,
the corresponding variable will filter out. By default is 1e-4.}

\item{id_out}{Whether output the id of variable kept in final output.}
}
\value{
A list of filtered data and/or index.
}
\description{
Filter the composition matrix
}
