% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNEMoE.R
\name{fitNEMoE}
\alias{fitNEMoE}
\title{Fit NEMoE model}
\usage{
fitNEMoE(
  NEMoE_obj,
  beta_init = NULL,
  gamma_init = NULL,
  num_restart = 5,
  restart_it = 10,
  BPPARAM = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{NEMoE_obj}{a NEMoE object contain data and parameters for fitting.}

\item{beta_init}{initial values of parameters in experts network.}

\item{gamma_init}{initial value of parameters in gating network.}

\item{num_restart}{Number of restart times.}

\item{restart_it}{Maximial number of iterations during restart.
the \code{BiocParallel} package is used. Default is SerialParam().}

\item{BPPARAM}{A \code{BiocParallelParam} class object from.}

\item{...}{other parameters can be passed to fitNEMoE.
See \code{\link{createParameterList}}}
}
\value{
a NEMoE object contain fitted result including parameters
beta a list contain coefficients of experts network in all levels.
gamma a matrix of coefficients in gating network.
r_i an array of inherit partition result in EM algorithm.
pen_fac estimated penalty factor for lambda1 (adapt = T). The (balanced)
lambda1 in NEMoE is lambda1*pen_fac
latent estimated latent classes.
}
\description{
This function fit NEMoE model using EM algorithm.
}
\examples{
data(NEMoE_example)
NEMoE_example
NEMoE_example <- fitNEMoE(NEMoE_example, num_restart = 0)
}
\seealso{
\code{\link{NEMoE_buildFromList}},
\code{\link{NEMoE_buildFromPhyloseq}}, \code{\link{createParameterList}}
}
