% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitNEMoE0}
\alias{fitNEMoE0}
\title{Cpp function: fitting NEMoE paramters.
(observed likelihood, penalized likelihood,
complete likelihood, penalized complete likelihood)
in all levels.}
\usage{
fitNEMoE0(
  X,
  seg,
  Z,
  y,
  K,
  lambda1,
  lambda2,
  alpha1,
  alpha2,
  V_init,
  W_init,
  beta_max,
  EM_opt,
  itmax,
  itmin,
  adapt,
  btr,
  stop_all,
  verbose,
  early_stop = FALSE
)
}
\arguments{
\item{X}{an aggregated data matrix (n*P) of input in experts network
(rbind of all levels input).}

\item{seg}{an integer vector of length in each level (L).}

\item{Z}{a data matrix (n*q) of input in gating network.}

\item{y}{a vector of response (n).}

\item{K}{A number of latent class.}

\item{lambda1}{aggregated penalty lambda parameters in experts network(P*K).}

\item{lambda2}{penalty parameters lambda in gating network(q*1).}

\item{alpha1}{penalty parameters alpha in experts network(L*K).}

\item{alpha2}{penalty parameter alpha in gating network(1).}

\item{V_init}{initial parameters in gating network((q+1)*K).}

\item{W_init}{aggregated initial parameters in experts network
(cbind of all levels) ((P+L)*K).}

\item{beta_max}{A number of maximal of coefficients to avoid divergence of
during the fitting.}

\item{EM_opt}{A integer indicate methods for EM algorithm.
0="EM", 1 = "CEM", 2 = "SEM", 3= "SAEM", 4="GEM".}

\item{itmax}{maximal numbers of iteration in fitting NEMoE.}

\item{itmin}{minimal numbers of iteration in fitting NEMoE.}

\item{adapt}{A boolean variable indicates whether to use adaptive NEMoE.}

\item{btr}{A boolean variable indicates whether to use backtracking during
fitting NEMoE.}

\item{stop_all}{A boolean variable indicates whether to stop by
(likelihood converge)&(parameters converge)}

\item{verbose}{A boolean variable indicates whether to show PLL
during each iteration.}

\item{early_stop}{A boolean variable indicates whether to stop when one
latent class have select zero variables (to save time).}
}
\value{
A matrix of fitting result.
}
\description{
Cpp function: fitting NEMoE paramters.
(observed likelihood, penalized likelihood,
complete likelihood, penalized complete likelihood)
in all levels.
}
