% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{getCoef}
\alias{getCoef}
\title{Get coefficients of fitted NEMoE object}
\usage{
getCoef(NEMoE_obj)
}
\arguments{
\item{NEMoE_obj}{A NEMoE object.}
}
\value{
A list contain fitted result.
}
\description{
This function get coefficients in NEMoE object.
}
\seealso{
\code{\link{fitNEMoE}}
}
