% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMoE_class.R
\docType{class}
\name{NEMoE-class}
\alias{NEMoE-class}
\alias{NEMoE}
\title{NEMoE class}
\description{
NEMoE class
}
\section{Slots}{

\describe{
\item{\code{Microbiome}}{A list of microbiome matrix, rows are samples.}

\item{\code{Nutrition}}{A matrix of nutrition intake.}

\item{\code{Response}}{A vector of health outcome.}

\item{\code{params}}{A list of parameters for fitting NEMoE.}

\item{\code{K}}{A number of fitting componenets in NEMoE (>=2).}

\item{\code{cvParams}}{A list of parameters for selecting parameters of NEMoE
using cross validation.}

\item{\code{cvResult}}{A dataframe of cross validation results of different parameters
of fitting NEMoE.}

\item{\code{NEMoE_output}}{A list NEMoE fitting result.}

\item{\code{taxLevel}}{A character indicates selected taxa level.}

\item{\code{ResponseLevel}}{A character indicates levels of response variables.}

\item{\code{standardize}}{A Logical variable indicate whether standardize input data.}
}}

