% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{NEMoE_buildFromList}
\alias{NEMoE_buildFromList}
\title{Build NEMoE object from list, dataframe or matrix}
\usage{
NEMoE_buildFromList(
  Microbiome,
  Nutrition,
  Response,
  K = NULL,
  cvParams = list(),
  taxLevel = NULL,
  taxTab = NULL,
  ...
)
}
\arguments{
\item{Microbiome}{A list of transformed microbiome
data of different taxonomic level. If it is a dataframe or matrix, will
convert to a list automatically.}

\item{Nutrition}{A dataframe of matrix of Nutrition matrix.}

\item{Response}{A numeric(factor) vector of Health outcome.}

\item{K}{A integer of number of latent class.}

\item{cvParams}{A list of cross validation parameters.}

\item{taxLevel}{A character of selected name of taxonomic levels.}

\item{taxTab}{A dataframe of taxonomic table.}

\item{...}{Other parameters can pass to NEMoE_buildFromList.
See \code{\link{createParameterList}}}
}
\value{
A NEMoE object.
}
\description{
This function build NEMoE object from List, dataframe or matrix.
}
\examples{
data(PD)
data_list = PD$data_list
Microbiome = data_list$Microbiom
Nutrition = data_list$Nutrition
Response = data_list$Response
NEMoE <- NEMoE_buildFromList(Microbiome = Microbiome, Nutrition = Nutrition, Response = Response)
}
