% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PD}
\alias{PD}
\title{16S gut microbiome Parkinson's disease dataset}
\format{
A list of two type of data:
\describe{
\item{ps}{A phyloseq object with 168 samples and 995 taxa}
\item{data_list}{A list of Microbiome, Nutrition and response variables}
}
}
\usage{
data(PD, package = 'NEMoE')
}
\description{
A list of two type of data (A phyloseq object and a list that can be put into NEMoE.
Microbiome data：
The phyloseq object contain 995 taxa and 168 samples.
The data list contain 168 samples with 4 levels(5, 16, 24, 36, 101) variables respectively.
Nutrition data: 168 samples and 27 variables.
PD state: PD = 1 and HC = 0.
}
\keyword{datasets}
