% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMoE_plot.R
\name{plotExperts}
\alias{plotExperts}
\title{Plot experts network}
\usage{
plotExperts(NEMoE_obj)
}
\arguments{
\item{NEMoE_obj}{a NEMoE object with fitted output.}
}
\value{
A list of graph of fitted coefficients of each latent class
on each level.
}
\description{
This function plot the estimated coefficients of each latent class
on each level.
}
\examples{
data(NEMoE_example)
plotExperts(NEMoE_example)
}
