% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMoE_plot.R
\name{plotGating}
\alias{plotGating}
\title{Plot gating network}
\usage{
plotGating(NEMoE_obj, PCs = c(1, 2))
}
\arguments{
\item{NEMoE_obj}{a NEMoE object with fitted output.}

\item{PCs}{Visualization of selected Principal components.}
}
\value{
A graph of PCA plot of nutrition intake and estimated latent classes
and its corresponding loadings.
}
\description{
This function plot the PCA of fitted latent class
and their corresponding loadings.
}
\examples{
data(NEMoE_example)
plotGating(NEMoE_example)
}
