% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rDirichlet}
\alias{rDirichlet}
\title{Cpp function: Sample from a Dirichlet distribution}
\usage{
rDirichlet(n, a)
}
\arguments{
\item{n}{Number of random vector to return.}

\item{a}{A positive vector of parameters.}
}
\value{
A random matrix with n row from Dirichlet
distribution with parameter a.
}
\description{
Cpp function: Sample from a Dirichlet distribution
}
