% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{setParam}
\alias{setParam}
\title{Set parameters of fitting NEMoE object}
\usage{
setParam(NEMoE_obj, ...)
}
\arguments{
\item{NEMoE_obj}{A NEMoE object.}

\item{...}{Other parameters can pass to setParam.}
}
\value{
A NEMoE with user input parameters.
}
\description{
This function set parameters in NEMoE object.
}
\examples{
data(NEMoE_example)
NEMoE_example <- setParam(NEMoE_example, lambda1 = 0.03)
}
\seealso{
\code{\link{createParameterList}}
}
