% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sMulti}
\alias{sMulti}
\title{Cpp function: fitting sparse multinomial regression.}
\usage{
sMulti(X, y, lambda, R, alpha, beta_max)
}
\arguments{
\item{X}{input matrix for sparse multinomial regression(n*p).}

\item{y}{response data for sparse multinomial regression(n*K).}

\item{lambda}{A vector of penalty parameters lambda (p*1).}

\item{R}{A vector of weighted parameters for each sample(n*1).}

\item{alpha}{A number of penalty parameters alpha.}

\item{beta_max}{A number of maximal number of fitted coefficients.}
}
\value{
A matrix of fitted coefficients.
}
\description{
Cpp function: fitting sparse multinomial regression.
}
